+++ test/deployment.nix
@@ -1,0 +1,10 @@
+combined: assertMsg:
+assert assertMsg
+  (combined.sortedDeployments == [
+    "firstArtifact"
+    "specificArtifact"
+    "anotherArtifact"
+    "lastArtifact"
+  ])
+  "Combined deployments did not have expected order.";
+{ }